ai-service/ai-service/requirements.txt
fastapi
uvicorn
keybert
sentence-transformers
transformers
torch==2.3.1
praw
numpy
scikit-learn
pydantic
python-dotenv
requests
mistralai
playwright
psycopg2-binary
supabase
openai
pandas
asyncpg
beautifulsoup4